Refactor out BalanceAndProxyAllowanceLazyStore
